move ui widgets out of home screen
